Perform es6 -> es5 transpilation post intall
Previously I was generating these files as a part of the build process
and checking them in. This makes each of the commits too noisy - the
changes to the map file are larger than the rest of the commit! Now I
generate the es5 and map files in post install. This is against the
guidance from the NPM docs, but the alternative appears to be worse.