Rework tests (part 1)
This change is a refactor to the way tests are written. It uses an
example filesystem to cut down on the initial set up. It also adds a
number of helpers to make the expectations simpler to express.

In this pass we have most of the "before ready" tests done and are
starting on "after ready". Unfortunately there is a bug uncovered by
some of the new "after ready" tests. The failing test and fix will go
into a subsequent commit.