Rework tests (part 2) + stat fix for "addDir"
In this commit we add a test for copying new directories from src to
dest after the "ready" event. It turns out that we do not get a stat
object (at least on OSX) and this in turn uncovered a bug where we stat
with the wrong path. This commit fixes this bug and also elaborates on
why we were not always getting stat objects from Chokidar - it turns out
this is expected behavior.