Handle timestamp preservation
Initially this was a somewhat novel port of the existing timestamp
tests. I discovered that the implementation did not work correctly when
trying to update a directory's mtime before hitting the "ready" state
because subsequent changes to the directory would again update its
mtime. Clearly we need to defer setting the mtime until after we've hit
the "ready" state.

This desire and the fact that I already had something else that needed
to wait for the ready state led me to create a list of "post ready
functions", which are evaluated immediately after getting the "ready"
event from Chokidar.

I also made a minor improvement by caching whether we're preserving file
and dir timestamps in the constructor.